Adding reusable modal component
