+++ src/AppRoutes.jsx
@@ -2,12 +2,12 @@
 import React from 'react';
 import { Routes as ReactRoutes, Route } from 'react-router-dom';
 import Stopwatch from './Components/Stopwatch';
-import ConfigurationPage from './Components/ConfigurationPage';
+import ConfigurationPanel from './Components/ConfigurationPanel';
 
 function AppRoutes() {
   return (
     <ReactRoutes>
-      <Route path="/configuration" element={<ConfigurationPage />} />
+      <Route path="/configuration" element={<ConfigurationPanel />} />
       <Route path="/" element={<Stopwatch />} />
     </ReactRoutes>
   );
