--- src/Components/ConfigurationPage/index.jsx
@@ -1,15 +1,0 @@
-// ConfigurationPage.js
-import React from 'react';
-import ConfigurationPanel from 'Components/ConfigurationPanel'; 
-
-function ConfigurationPage() {
-  return (
-    <div>
-      {/* Content for your configuration page */}
-      <h1>Configuration Page</h1>
-      <ConfigurationPanel />
-    </div>
-  );
-}
-
-export default ConfigurationPage;
