+++ src/Components/ConfigurationPanel/ConfigurationPanel.js
@@ -1,15 +1,15 @@
+// ConfigurationContent.jsx
 import './index.css'
 
 import TimerOption from 'Components/TimerOption'
+import React from 'react';
 
-const ConfigurationPanel = () => {
-    return (
-        <div className='configuration-panel-wrapper'>
-            <TimerOption icon={'duration-icon'} label={'Duration'} valueSelectorType={'time'}/>
-            <TimerOption icon={'sound-icon'} label={'Sound'} valueSelectorType={'sound'}/>
-            <TimerOption icon={'rest-icon'} label={'Rest Break'} valueSelectorType={'dropdown'}/>
-        </div>
-    )
-}
+const ConfigurationPanel = ({ onClose }) => (
+    <div className='configuration-panel-wrapper'>
+    <TimerOption icon={'duration-icon'} label={'Duration'} valueSelectorType={'time'} />
+    <TimerOption icon={'sound-icon'} label={'Sound'} valueSelectorType={'sound'} />
+    <TimerOption icon={'rest-icon'} label={'Rest Break'} valueSelectorType={'dropdown'} />
+</div>
+);
 
-export default ConfigurationPanel
+export default ConfigurationPanel;
