+++ src/Components/NavBar/index.js
@@ -1,0 +1,2 @@
+// import * from './NavBar'
+export { default } from './NavBar'
