+++ src/Components/ReusableModal/ReusableModal.css
@@ -1,0 +1,20 @@
+.modal-dark .modal-content {
+    background-color: #343a40;
+    color: #fff;
+}
+
+.modal-dark .modal-header {
+    border-bottom: 1px solid #495057;
+}
+
+.modal-dark .modal-header button.close {
+    color: #fff;
+}
+
+.modal-dark .modal-body {
+    color: #fff;
+}
+
+.modal-dark .modal-footer {
+    border-top: 1px solid #495057;
+}
