+++ src/Components/ReusableModal/index.js
@@ -1,0 +1,2 @@
+// import * from './ReusableModal'
+export { default } from './ReusableModal'
